refactor: restructure package layout, normalize imports, and fix scheduler bugs
## Package layout
Replaced the internal/pkg split with a flat, idiomatic layout:
  internal/config           → config/
  internal/errors           → errors/
  internal/task             → models/
  internal/repository/      → repositories/
  internal/service/         → services/
  internal/transport/       → http/
  internal/validate/        → utils/helpers/
  internal/version/         → utils/version/
  pkg/notifier/             → utils/notifications/
  pkg/timex/ + pkg/util/    → utils/helpers/
  pkg/httpclient/           → utils/httpclient/

Deleted internal/config/loader.go and internal/logger/logger.go, whose
responsibilities were absorbed into main.go and config/.

## Imports
All files now group imports into three labelled sections:
  // Go Internal Packages
  // Local Packages        (explicit alias = package name on every entry)
  // External Packages

## Validation helpers
Added helpers.ValidateDate and helpers.ValidateTime to utils/helpers/validate.go.
models.CreateRequest.Validate() now delegates date/time required+format checks
to these helpers, removing duplicate inline logic.

## Bug fixes
- scheduleTaskNow: held tasksMu across the exists-check, cron.AddJob, and
  s.tasks map-write atomically, eliminating a TOCTOU race that could
  double-schedule a task and leak an un-removable cron entry.
- scheduleExistingTask: added intervalInSeconds == 0 guard before the modulo
  to prevent a divide-by-zero panic on corrupt recurring tasks (recur = 0).
- SchedulerService: added Stop() that calls cron.Stop() then execCancel().
  closeCallback now calls Stop() before mongoClient.Close() so cron jobs
  cannot fire against a closed database connection during graceful shutdown.
- Executor context: replaced the stored server-lifecycle s.ctx with a
  dedicated execCtx/execCancel owned by SchedulerService. Executors now
  survive the shutdown signal and are cancelled only by Stop(), preventing
  abrupt mid-retry cancellations.
- executer.Run: replaced time.After with time.NewTimer and an explicit
  timer.Stop() in the ctx.Done branch to eliminate a goroutine leak during
  retry backoff.
- http/server.go: restored the HealthChecker interface (removed concrete
  *health.HealthCheckService dependency from the HTTP layer).
- main.go: fixed log.Fatal(..., zap.Error(err)) → logger.Fatal(...) so
  startup errors are logged in structured logfmt instead of a raw struct.
