+++ .gitignore
@@ -1,10 +1,70 @@
-.idea
-.bin
+# These are some examples of commonly ignored file patterns.
+# You should customize this list as applicable to your project.
+# Learn more about .gitignore:
+#     https://www.atlassian.com/git/tutorials/saving-changes/gitignore
+
+# Node artifact files
+node_modules/
+dist/
+
+# Compiled Java class files
+*.class
+
+# Compiled Python bytecode
+*.py[cod]
+
+# Log files
+*.log
+
+# Package files
+*.jar
+
+# Maven
+target/
+
+bin/
+.bin/
+
+# Unit test reports
+TEST*.xml
+
+# Generated by MacOS
 .DS_Store
-venv
-.venv
-/.claude/
+
+# Generated by Windows
+Thumbs.db
+
+# Applications
+*.app
+*.exe
+*.war
+
+# Large media files
+*.mp4
+*.tiff
+*.avi
+*.flv
+*.mov
+*.wmv
+
+# Config files
 /config/*.yaml
 /config/*.yml
+
+# Docker related
+.docker/
+
+# JetBrains IDE
+.idea/
+
+# VSCode related
+.vscode/
+
+# Cursor related
+.cursor/
+
+# claude
+/.claude/
+
 coverage.out
 coverage.html
