+++ deployments/docker-compose.yml
@@ -5,11 +5,6 @@ services:
     network_mode: host
     volumes:
       - mongo_data:/data/db
-    healthcheck:
-      test: ["CMD", "mongosh", "--eval", "db.adminCommand('ping')"]
-      interval: 10s
-      timeout: 5s
-      retries: 5
 
 volumes:
   mongo_data:
