+++ go.mod
@@ -1,6 +1,6 @@
 module scheduler
 
-go 1.25.0
+go 1.26.1
 
 require (
 	github.com/alecthomas/kingpin/v2 v2.4.0
