--- pkg/notifier/notifier.go
@@ -1,11 +1,0 @@
-package notifier
-
-import (
-	"context"
-
-	task "scheduler/internal/task"
-)
-
-type Sender interface {
-	SendAlert(ctx context.Context, t task.Task, errMsg string) error
-}
